requirements/base.txt
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
billiard==3.3.0.23
boto3==1.4.4
botocore==1.5.69
celery==3.1.25
cffi==1.10.0
coreapi==2.3.0
coreschema==0.0.4
croniter==0.3.16
decorator==4.0.11
defusedxml==0.5.0
Django==1.11
django-appconf==1.0.2
django-celery==3.2.1
django-cors-headers==2.1.0
django-datatables-view==1.14.0
django-environ==0.4.3
django-filter==1.0.2
django-fsm==2.5.0
django-fsm-admin==1.2.4
django-fsm-log==1.3.0
django-redis==4.7.0
django-rest-swagger==2.1.2
django-silk==1.0.0
django-sql-explorer==1.1.1
djangorestframework==3.6.2
djangorestframework-expiring-authtoken==0.1.4
djangorestframework-gis==0.11
djangorestframework-xml==1.3.0
docutils==0.13.1
fcm-django==0.2.7
humanize==0.5.1
ipython==5.3.0
ipython-genutils==0.2.0
itypes==1.1.0
Jinja2==2.9.6
jmespath==0.9.3
kombu==3.0.37
Markdown==2.6.8
MarkupSafe==1.0
openapi-codec==1.3.1
pexpect==4.2.1
phonenumbers==8.4.2
pickleshare==0.7.4
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
pycparser==2.17
pycrypto==2.6.1
pyfcm==1.3.1
Pygments==2.2.0
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
redis==2.10.5
requests==2.13.0
requests-toolbelt==0.8.0
s3transfer==0.1.10
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
sqlparse==0.2.3
traitlets==4.3.2
unicodecsv==0.14.1
uritemplate==3.0.0
wcwidth==0.1.7
XlsxWriter==0.9.6
newrelic
django-health-check
https://s3-eu-west-1.amazonaws.com/fetchr.pypi/fetchr_api_response-0.1.5.tar.gz
social-auth-app-django==1.2.0
coverage==4.4.1
ipdb==0.10.2
